Create separate file for g2p stable id
